compiler/ir2tran: Disable most of the important bits on ARM.
  * This combination of outright disabling functions and replacing
various calls to VOP with calls to ERROR, conditionalized for ARM,
reduces the set of calls to VOP to just those to
XEP-ALLOCATE-FRAME, SETUP-ENVIRONMENT, and RETURN-SINGLE.

  * This combination of VOPs should, in theory, be sufficient to
compile a leaf function that takes a fixed number of arguments,
does not use a closure environment, and returns only a single
value.  While such a function can be of practical use, a compiler
that can compile only such functions cannot.  The upside here is
that such a compiler should be easy to debug, and can be extended
towards a more useful subset afterwards.