compiler/arm/insts: Add an "instruction" for finding an LRA during call.
  * Having an instruction to find the location of an LRA (or CODE,
for that matter) is a fairly normal thing for a backend to have.

  * In our case, we need to find the address for an LRA given the
location of the LRA header.  This is actually doable as an ADD, so
long as it's done once we know the relative locations of both the
ADD instruction itself and the LRA header, for which we need a
back-patch.