compiler/arm/parms: Define more static symbols.
  * The runtime believes that any backend that doesn't define
reg_ALLOC has a static symbol called *ALLOCATION-POINTER*.  Until
now that has been x86oids only (due to register paucity on the
former and inertia on the latter), but it's a perfectly reasonable
thing to do when you have less than thirty-two registers available.

  * While we're here, also define symbols for the binding stack
pointer (another x86oidism) and separate symbols for pseudo-atomic
and pseudo-atomic-interrupted.

  * The x86oids give us precedent for P-A in a symbol, but they
also have atomic R-M-W memory instructions.  We might be able to
do something on the order of an atomic compare-exchange, but it's
simply easier to separate the P-A bits into two separate symbols,
declare them to be generalized booleans (since we can compare to
NIL easily and have several registers that will contain something
that can serve as a FIXNUM).