compiler/arm: Macro LISP-RETURN should adjust flags, not the call sites.
  * Rather than have the call sites do direct manipulation of the
CPSR, have them pass a parameter to LISP-RETURN and let it do the
CPSR manipulation.  This brings the details of how a multi-valued
return is signalled into one point on the returning side, even if
it is still distributed on the receiving side.  It also more
closely parallels the other backends, although they traditionally
expose the details more than we do here.