compiler/arm/macros: Error breakpoint emission.
  * Based on the PPC backend.

  * Unlike any other backend, ARM requires the use of a register
to hold the syscall number for a break instruction.  This wasn't
originally the case, but since the move from OABI to EABI it has
become necessary.  Should we end up wanting to support OABI as a
build option, we can (must!) pass the syscall number as an
immediate constant in the SWI instruction, but that's unlikely at
this point.