compiler/arm/macros,runtime/arm-arch.c: PSEUDO-ATOMIC.
  * Define PSEUDO-ATOMIC macro based loosely on the PPC backend,
altered to suit the realities of having both *PSEUDO-ATOMIC-ATOMIC*
and *PSEUDO-ATOMIC-INTERRUPTED* in registers, and a scheme for
using a conditional breakpoint for the pending-interrupt trap.

  * Alter arch_set_pseudo_atomic_interrupted() to suit our new
conditional breakpoint scheme by storing the syscall number for a
breakpoint as the "interrupted" value.

  * Alter sigtrap_handler() to make sure that the current
instruction is an SWI, and to dispatch based on its condition code.

  * Set up the ALWAYS condition (AL) to process a trap as normal.

  * Set up the LESS-THAN condition (LT) to process a pending-
interrupt trap.

  * Set up the remaining conditions to lose() for now.  We may
have use for them later, but we should fail hard if we see them in
the meantime.