ARM PSEUDO-ATOMIC fixes.
  * At several points in PSEUDO-ATOMIC I neglected to unquote the
various inputs to the macro, leading to build failures when the
macro was used.

  * The static symbols used for PSEUDO-ATOMIC were not being
initialized, remaining unbound by default, leading to the runtime
thinking that the system is in P-A until the end of the first P-A
block...  And then executing a syscall (the number of which is
(ASH UNBOUND-MARKER-WIDETAG (- N-FIXNUM-TAG-BITS)), and that I've
never bothered to look up) which clobbers at least R0 and possibly
other registers.  Fixed, by initializing both to NIL during
startup if *PSEUDO-ATOMIC-INTERRUPTED* is a static symbol.