compiler/: Make ARM OCFP-handling more like x86oid OCFP-handling.
  * Continuing our saga of simplifying the ARM port changes by
moving more towards an x86oid frame-pointer handling model, we
reinstate the VM support routine for the OCFP passing location, as
it's used in precisely one place, and is handed off to a MOVE VOP
that uses LOCATION= to decide if it actually needs to do anything.
If it doesn't then it emits no code, and making sure it doesn't is
trivial.

  * Crib the new VM support routine definition from the x86
backend, as it already does what we need here.

  * Remove the reader conditional that disables the MOVE in
question, now that the MOVE is guaranteed to do no harm.