compiler/arm/move: Rough in remaining parts of the file.
  * Cribbed entirely from the PPC backend, with some fairly major
adaptation required for the potentially-bignum-boxing
MOVE-FROM-SIGNED and MOVE-FROM-UNSIGNED VOPs.  Also largely
untested.

  * There's no way we're lucky enough to not have to touch this
file again, but this should get the worst of the nearly
incomprehensible errors to do with such-and-such TN is not valid
for such-and-such purpose because blah blah blah SCs blah blah.
I'm so bloody tired of trying to figure out which tiny fragment of
the system is missing when that happens, and there was little
enough left to do in this file, so I figured I may as well just
finish the rough-in and move on to other things.