compiler/arm/{move,vm}: Allow more immediate-constant values.
  * Apparently, the compiler Really Likes to be able to use a
FIXNUM as an immediate constant.  As in, it can't load FIXNUMs
from the code component because the SCs aren't compatible or
something.

  * Also apparently, I didn't know what I was doing when first I
set up the VM definition and IMMEDIATE-CONSTANT-SC, because for
some reason I seem to be restricting things to what I can load
from scratch in a single instruction, which simply isn't plausible.

  * So, expand IMMEDIATE-CONSTANT-SC to accept the entire FIXNUM
range (implemented in terms of INTEGER and the XC constants as
opposed to FIXNUM because the TYPECASE is evaluated on the host,
which might (definitely does, in my case) have a different idea of
how large a FIXNUM is) and CHARACTERs while we're at it.

  * And hack something in to LOAD-IMMEDIATE so that it can handle
both the full range of FIXNUMs and the full range of characters,
even if it doesn't produce the most wonderfully compact code for
all possible inputs...  Or, it turns out, correct code for
negative FIXNUMs.  I'll have to come back to that.