compiler/arm/insts: New pseudo-instruction, LOAD-FROM-LABEL.
  * When we deal with fixups, which have to be stored in memory as
a cohesive word-wide value, we tend to store them in the ELSEWHERE
segment and use LDR of a label to retreive them.  This does not
work well when the code component exceeds a certain size.

  * What would work well is a multi-instruction sequece, building
a displaced address in an interior-pointer register until the
difference between the position of the label and the address built
is within the range of an LDR instruction.

  * So do precisely that, in one place, with choosers and whatnot
to optimize for a small delta as best we can.