compiler/arm/c-call: Relieve register pressure during CALL-OUT.
  * Several things wrong here, including a previously unexplored
constraint.  We have an NFP which is component-live, up to four
C arguments passed in registers, and only five unboxed registers
to work with...  And a temporary bound to an unboxed register and
the function to call being passed in a SAP-REG which is also
unboxed.  That makes seven pounds of s...tuff in a five pound bag.

  * First, the temporary is used to hold the address of
call_into_c, which is word-aligned, and thus looks like a FIXNUM
to the GC.  Thus, it can go in an ANY-REG.

  * Second, about the only place that we can put FUNCTION that
will be acceptable to the compiler (due to primtype restrictions)
and that we can get at is the number stack, so force it to
SAP-STACK instead of SAP-REG, and use a LOADW (cribbed from
LOAD-NUMBER-STACK) to get at it.

  * This brings our register pressure on CALL-OUT down about as
low as it can get, and it should be low enough to cover the
maximum number of C register arguments.