Implement DEFINE-LOAD-TIME-GLOBAL.
This is the patch proposed on sbcl-devel (ca. November 2013)
with alterations for 2 out of 3 concerns raised by Nikodemus.

Indeed users can not roll their own equivalent of the new macro
using just the new proclamation. It's confusing to expose the
feature, as it would mean introducing another thing like
  (DECLAIM (EVENTUALLY-ALWAYS-BOUND X)) ; *HYPOTHETICAL*
and then explaining how compilation "times" a/k/a "situations" are
not quite the same as "now" and "eventually", and why 2N choices
exist for one concept. YAGNI/Occam's Razor come to mind.

Were a user to try something like
  (eval-when (:load-toplevel) (proclaim '(always-bound x)))
it can not possibly have any unusual effect due to PROCLAIM being
an ordinary function, whereas eval-when :compile-toplevel of same
certainly means that X is right now always bound.
So only the new macro is documented, not its implementation.

Closes lp#1253688