compiler/ir1opt: Don't break STACK analysis in MAYBE-DELETE-EXIT.
  * When deleting a local exit that is delivering multiple-valued
results to its continuation, and there is another live exit path
that could deliver already-computed multiple-valued results to the
same continuation, simply using DELETE-FILTER to unlink the EXIT
node and substitute the LVARs causes stack analysis to believe
that the previous value of the LVAR is dead well prior to the
(being-unlinked) EXIT node.

  * Instead, what we need to do is to splice in a new block that
contains a CAST node to overwrite the continuation LVAR, and to
arrange it so that the CAST cannot be merged back onto the block
that contained the EXIT, allowing STACK analysis to conclude,
correctly, that the previous value is still live until after the
block formerly containing the EXIT node.

  * This is bugs lp#309099, lp#518099 and lp#533930, fixed at long
last.