Properly pprint literal functions inside nested lists.
SB-IMPL::OUTPUT-FUN used to bind *print-level* to 3, causing
`(((,#'eq))) to be printed as (((#))), it explained that that was for
printing of interpreted functions, but interpreted functions aren't
printed as code anymore, just a name.
That *print-level* is now removed, although it could be solved by
binding it to (+ *current-level-in-print* 3), if required.

Fixes lp#1300716.