Shrink symbols on x86-64 by putting tls-index in the header.
Using the upper 4 bytes of the header makes symbols take up
6 words instead of 8, reducing the initial core by 565KB.

The patch could theoretically be applied for 32-bit builds,
but using a 2-byte index limits the TLS_SIZE to 2^14 words.
This is sufficient at present, but I wouldn't want to undo
the space savings if it is decided that TLS_SIZE should
be enlarged by a factor of 8 or more in the future.

In C, tls_index_of() should be used in lieu of sym->tls_index.
This is defined to be the constant 0 for non-threaded builds,
so eliminates #ifdef LISP_FEATURE_SB_THREAD in three places.