compiler/arm/{insts,vm}: Fix up float register damage.
  * First, the register spaces overlap, meaning that they must
share an SB, and thus also the encoding for the offset space.  Fix
to share FLOAT-REGISTERS instead of having separate single and
double float register files.  Update insts to deal with the newer
encoding for double-float offsets (removing the "precision" noise).

  * Second, we've just collapsed a couple of SBs into one, so we
need to update LOCATION-PRINT-NAME.  Rather than saying S~D or
D~D, we just say F~D like on all of the other platforms (we could
get clever the way that x86oids do with the SC name, but that can
easily wait until later).

  * Third, the comments in vm.lisp about :element-size not being
valid for LOAD-TNs are just a little bit out of date.  The comment
was originally written for the PPC backend in early February 1991,
based on a similar comment in the MIPS backend from the previous
September.  PACK was updated to honor :element-size for SCs in
late February 1991.  Remove the comments.

  * And now that we have the SBs sorted out, we update the SCs to
use appropriate offsets within the SB and element sizes, following
the SPARC precedent, where they actually got it somewhat right.