compiler/arm/{arith,parms}: Define remaining two-arg static-funs.
  * These are all of the remaining TWO-ARG-whatever static-funs
actually implemented in lisp code.  Presumably they are also the
minimal required set, but in any case defining more would be
overkill.

  * While this is not the last of the static-funs, I suspect that
it is the last of the REQUIRED static-funs, as the remaining cases
all :TRANSLATE the function name (as opposed to these cases, where
the :TRANSLATE is for a different base function).