Move ARM control stack pointer to a static-symbol.
  * The simplest part of this whole mess is in src/compiler/arm/,
where we define a couple of new static-symbols (one for the
runtime to use during interrupt processing) and we switch out the
LOAD-CSP and STORE-CSP macros in favor of using a symbol value.

  * The next tricky bit is in interrupt.c, where we change how we
access the control stack pointer on ARM, and where we arrange to
store the control stack pointer where a non-local unwind won't
screw up if it runs across it but still where it can be restored
if we do a "normal" exit (necessary because it's no longer part of
the signal context).

  * And the really tricky bit is fixing up call_into_lisp and
call_into_c to operate in terms of the new symbol instead of in
terms of the old register.  Hopefully I got that right, and it's
about to change again, anyway.