Remove ~200KB of uninterned symbol names from the image.
First, DEFMACRO always names its WHOLE and ENVIRONMENT arguments the same
by STRING= but using non-EQ symbols. This is subtly different from
(LET ((whole '#:whole) (environment #':env)) ..) - because if for some
reason that eludes me there is really an issue of inadvertent capture,
the compile-time hygiene of macros is essentially unaltered by this patch.

Second, genesis coalesces uninterned symbols whose names are STRING=.
This is permissible not just for our own code, but I refrained from having
the regular fasloader do this because I suspect that it might be suprising
that uninterned symbols are coalesced across files. It's hard to tell what
CLHS had in mind- it doesn't specify when objects have to be *dissimilar*.